Move hl.grid tests to their own file
stack-info: PR: https://github.com/pytorch-labs/helion/pull/208, branch: jansel/stack/59
